Fix mixed active content error
whateverorigin.org and similar sites use http, rather than https. As
such, it works locally when hosted on http://, but breaks when running
on github pages. We're switching the CORS bypass mechanism in this
commit to one hosted on https to fix this error.

A side effect of this is that we don't need jQuery anymore.